+++ CHANGELOG.md
@@ -1,4 +1,26 @@
 # Changelog
+### 0.8.24
+  - cancelAll({ resetState: true }) can remove existing derived state (#253)
+  - fix issue with Ember 3.6.0+ when IE11 is not being targeted (#261)
+
+### 0.8.22
+  - Fixes ember-metal related issues on Ember Canary.
+
+### 0.8.21
+  - Fixes ES5 getter syntax on Ember Canary (#248)
+
+### 0.8.18
+  - Adds `.evented()` task modifier and task lifecycle events.
+
+### 0.8.17
+  - waitForEvent/Queue/Property helpers now work with the various
+    promise helpers (e.g. `all` and `race`)
+
+### 0.8.16
+  - Allow passing a non-Function value to waitForProperty() as
+    a shortcut for waiting for the observed property to `===`
+    that value.
+
 ### 0.8.15
   - Added waitForProperty() to pause execution until a property
     on an Ember Object becomes a certain value.